+++ Packages/com.unity.inputsystem/InputSystem/Actions/Composites/AxisComposite.cs
@@ -219,13 +219,12 @@ namespace UnityEngine.InputSystem.Composites
 
         public override void OnGUI()
         {
-#if UNITY_INPUT_SYSTEM_PROJECT_WIDE_ACTIONS
-            if (!InputSystem.settings.useIMGUIEditorForAssets) return;
-#endif
+            if (!InputSystem.settings.useIMGUIEditorForAssets)
+                return;
+
             target.whichSideWins = (AxisComposite.WhichSideWins)EditorGUILayout.EnumPopup(m_WhichAxisWinsLabel, target.whichSideWins);
         }
 
-#if UNITY_INPUT_SYSTEM_PROJECT_WIDE_ACTIONS
         public override void OnDrawVisualElements(VisualElement root, Action onChangedCallback)
         {
             var modeField = new EnumField(m_WhichAxisWinsLabel.text, target.whichSideWins)
@@ -241,8 +240,6 @@ namespace UnityEngine.InputSystem.Composites
 
             root.Add(modeField);
         }
-
-#endif
     }
     #endif
 }